feat(bootstrap4-theme): move the component to another folder
