+++ packages/bootstrap4-theme/src/scss/bootstrap-asu-extends.scss
@@ -42,3 +42,4 @@
 @import 'extends/tables';
 @import 'extends/person-profile';
 @import 'extends/image-parallax';
+@import 'extends/image-based-card';
