+++ packages/bootstrap4-theme/stories/molecules/cards/cards.examples.stories.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { createComponent, createStory } from '../../../helpers/wrapper.js';
 import exampleImage from './example-image.jpg';
-import portraitImage from './portrait-image.jpg';
 
 export default createComponent('Cards', 'Molecules', 'Examples');
 
@@ -1556,84 +1555,4 @@ export const mediaTypeOverlayCard = createStory(
     initFunc,
     omitTemplate: true,
   }
-);
-
-export const ImageBasedCardsWithCTA = createStory(
-  <>
-    <br />
-    <div class="content-section text-white">
-      <div class="image-holder">
-        <img
-          src={portraitImage}
-          alt="alt text"
-          loading="lazy"
-          decoding="async"
-        />
-      </div>
-      <div class="content-holder">
-        <div className="content-bg">
-          <h3>Serving all learners at every stage of life</h3>
-          <div class="hidden-details">
-            <div class="long-text">
-              As a comprehensive public research university, ASU is committed to
-              providing excellence in education through the Academic Enterprise,
-              and enables the success of each unique student and increases
-              access to higher education for all.
-            </div>
-            <a
-              href="#"
-              class="btn btn-gold btn-sm"
-              data-ga="Call to action"
-              data-ga-name="onclick"
-              data-ga-event="link"
-              data-ga-action="click"
-              data-ga-type="internal link"
-              data-ga-region="main content"
-              data-ga-secion="the new american university"
-            >
-              Call to Action
-            </a>
-          </div>
-        </div>
-      </div>
-    </div>
-    <br />
-    <div class="content-section text-white">
-      <div class="image-holder">
-        <img
-          src={portraitImage}
-          alt="alt text"
-          loading="lazy"
-          decoding="async"
-        />
-      </div>
-      <div class="content-holder">
-        <div className="content-bg">
-          <h3>Experience world-class</h3>
-          <div class="hidden-details">
-            <div class="long-text">
-              As a comprehensive public research university, ASU is committed to
-              providing excellence in education through the Academic Enterprise,
-              and enables the success of each unique student and increases
-              access to higher education for all.
-            </div>
-            <a
-              href="#"
-              class="btn btn-gold btn-sm"
-              data-ga="Call to action"
-              data-ga-name="onclick"
-              data-ga-event="link"
-              data-ga-action="click"
-              data-ga-type="internal link"
-              data-ga-region="main content"
-              data-ga-secion="the new american university"
-            >
-              Learn more
-            </a>
-          </div>
-        </div>
-      </div>
-    </div>
-    <br />
-  </>
 );
