+++ packages/bootstrap4-theme/stories/molecules/image-based-cards/portrait-image.jpg
